+++ Source/OxyPlot.Wpf/Annotations/PolylineAnnotation.cs
@@ -19,12 +19,6 @@ namespace OxyPlot.Wpf
     public class PolylineAnnotation : PathAnnotation
     {
         /// <summary>
-        /// Identifies the <see cref="Smooth"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty SmoothProperty = DependencyProperty.Register(
-            "Smooth", typeof(bool), typeof(PolylineAnnotation), new PropertyMetadata(false, DataChanged));
-
-        /// <summary>
         /// Identifies the <see cref="InterpolationAlgorithm"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty InterpolationAlgorithmProperty = DependencyProperty.Register(
@@ -51,7 +45,7 @@ namespace OxyPlot.Wpf
             TextHorizontalAlignmentProperty.OverrideMetadata(typeof(PolylineAnnotation), new FrameworkPropertyMetadata(HorizontalAlignment.Right, AppearanceChanged));
             TextVerticalAlignmentProperty.OverrideMetadata(typeof(PolylineAnnotation), new FrameworkPropertyMetadata(VerticalAlignment.Top, AppearanceChanged));
         }
-        
+
         /// <summary>
         /// Initializes a new instance of the <see cref="PolylineAnnotation" /> class.
         /// </summary>
@@ -90,20 +84,10 @@ namespace OxyPlot.Wpf
         }
 
         /// <summary>
-        /// Gets or sets a value indicating whether the polyline should be smoothed.
-        /// </summary>
-        /// <value><c>true</c> if smooth; otherwise, <c>false</c>.</value>
-        public bool Smooth
-        {
-            get { return (bool)this.GetValue(SmoothProperty); }
-            set { this.SetValue(SmoothProperty, value); }
-        }
-        
-        /// <summary>
-        /// Gets or sets a interpolation algorithm that should be used for smoothing.
+        /// Gets or sets the interpolation algorithm.
         /// </summary>
         /// <value>Interpolation algorithm.</value>
-        public IInterpolationAlgorithm InterpolationAlgorithm 
+        public IInterpolationAlgorithm InterpolationAlgorithm
         {
             get { return (IInterpolationAlgorithm)this.GetValue(InterpolationAlgorithmProperty); }
             set { this.SetValue(InterpolationAlgorithmProperty, value); }
@@ -129,13 +113,7 @@ namespace OxyPlot.Wpf
             var a = (Annotations.PolylineAnnotation)this.InternalAnnotation;
             a.Points.Clear();
             a.Points.AddRange(this.Points);
-
-            if (this.InterpolationAlgorithm == null && this.Smooth) {
-                a.InterpolationAlgorithm = InterpolationAlgorithms.CanonicalSpline;
-            } else {
-                a.InterpolationAlgorithm = this.InterpolationAlgorithm;
-            }
-
+            a.InterpolationAlgorithm = this.InterpolationAlgorithm;
             a.MinimumSegmentLength = this.MinimumSegmentLength;
         }
     }