+++ Source/OxyPlot.Wpf/Series/LineSeries.cs
@@ -146,12 +146,6 @@ namespace OxyPlot.Wpf
                 "MinimumSegmentLength", typeof(double), typeof(LineSeries), new PropertyMetadata(2.0, AppearanceChanged));
 
         /// <summary>
-        /// Identifies the <see cref="Smooth"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty SmoothProperty = DependencyProperty.Register(
-            "Smooth", typeof(bool), typeof(LineSeries), new UIPropertyMetadata(false));
-
-        /// <summary>
         /// Identifies the <see cref="InterpolationAlgorithm"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty InterpolationAlgorithmProperty = DependencyProperty.Register(
@@ -461,27 +455,10 @@ namespace OxyPlot.Wpf
         }
 
         /// <summary>
-        /// Gets or sets a value indicating whether this <see cref="DataPointSeries" /> is smooth.
-        /// </summary>
-        /// <value><c>true</c> if smooth; otherwise, <c>false</c> .</value>
-        public bool Smooth
-        {
-            get
-            {
-                return (bool)this.GetValue(SmoothProperty);
-            }
-
-            set
-            {
-                this.SetValue(SmoothProperty, value);
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets a value indicating what interpolation algorithm should be used for smoothing.
+        /// Gets or sets a value the interpolation algorithm.
         /// </summary>
         /// <value>Interpolation algorithm.</value>
-        public IInterpolationAlgorithm InterpolationAlgorithm 
+        public IInterpolationAlgorithm InterpolationAlgorithm
         {
             get
             {
@@ -548,12 +525,7 @@ namespace OxyPlot.Wpf
             s.BrokenLineStyle = this.BrokenLineStyle;
             s.BrokenLineThickness = this.BrokenLineThickness;
             s.Decimator = this.Decimator;
-
-            if (this.InterpolationAlgorithm == null && this.Smooth) {
-                s.InterpolationAlgorithm = InterpolationAlgorithms.CanonicalSpline;
-            } else {
-                s.InterpolationAlgorithm = this.InterpolationAlgorithm;
-            }
+            s.InterpolationAlgorithm = this.InterpolationAlgorithm;
         }
     }
 }