+++ Source/OxyPlot/Rendering/InterpolationAlgorithms.cs
@@ -1,4 +1,13 @@
-﻿namespace OxyPlot
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="InterpolationAlgorithms.cs" company="OxyPlot">
+//   Copyright (c) 2014 OxyPlot contributors
+// </copyright>
+// <summary>
+//   Implements a set of predefined interpolation algorithms.
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace OxyPlot
 {
     /// <summary>
     /// Implements a set of predefined interpolation algorithms.