+++ Source/OxyPlot/Series/LineSeries.cs
@@ -203,16 +203,6 @@ namespace OxyPlot.Series
         public double MinimumSegmentLength { get; set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether this <see cref = "DataPointSeries" /> is smooth.
-        /// </summary>
-        /// <value><c>true</c> if smooth; otherwise, <c>false</c>.</value>
-        [Obsolete("Use the InterpolationAlgorithm property instead")]
-        public bool Smooth {
-            get { return this.InterpolationAlgorithm is CanonicalSpline; }
-            set { this.InterpolationAlgorithm = value ? InterpolationAlgorithms.CanonicalSpline : null; }
-        }
-
-        /// <summary>
         /// Gets or sets a type of interpolation algorithm used for smoothing this <see cref = "DataPointSeries" />.
         /// </summary>
         /// <value>Type of interpolation algorithm.</value>
@@ -727,7 +717,7 @@ namespace OxyPlot.Series
             {
                 // spline smoothing (should only be used on small datasets)
                 var resampledPoints = ScreenPointHelper.ResamplePoints(pointsToRender, this.MinimumSegmentLength);
-                screenPoints = InterpolationAlgorithm.CreateSpline(resampledPoints, false, 0.25);
+                screenPoints = this.InterpolationAlgorithm.CreateSpline(resampledPoints, false, 0.25);
             }
 
             // clip the line segments with the clipping rectangle
@@ -768,7 +758,7 @@ namespace OxyPlot.Series
         protected virtual void ResetSmoothedPoints()
         {
             double tolerance = Math.Abs(Math.Max(this.MaxX - this.MinX, this.MaxY - this.MinY) / ToleranceDivisor);
-            this.smoothedPoints = InterpolationAlgorithm.CreateSpline(this.ActualPoints, false, tolerance);
+            this.smoothedPoints = this.InterpolationAlgorithm.CreateSpline(this.ActualPoints, false, tolerance);
         }
 
         /// <summary>